nosoop and caxanga334 merged

+++ utils/RCBot2_meta/bot_mtrand.cpp
@@ -97,7 +97,7 @@ int randomInt ( int imin, int imax )
 	// quicker
 	if (imin == imax)
 		return imin;
-	return static_cast<int>(imin + (int)(randomOne()*((float)(imax - imin) + 0.99f)));
+	return imin + (int)(randomOne()*((float)(imax - imin) + 0.99f));
 	//return (int)(imin + (int)(drand()*((float)(imax-imin)+0.99f)));
 }
